feat: add chat panel on workflow designer
